feat: move ReasonType to public
